Fix 22 build errors: casts, unused vars, function calls
ROCmGPUProbe fixes (15 errors):
- Add [[maybe_unused]] to RSMI_STATUS_* constants (lines 21-35)

ProcessDetailsPanel fixes (4 errors):
- Cast uint64→double for formatBytes() calls (lines 1102, 1193, 1302)
- Fix hoveredIndexFromPlotX: pass vector not pointer (lines 1256, 1296)

SystemMetricsPanel fixes (2 errors):
- Cast uint64→double for formatBytes() calls (lines 1500-1501)

All errors resolved, build should succeed

+++ src/App/Panels/SystemMetricsPanel.cpp
@@ -1497,8 +1497,8 @@ void SystemMetricsPanel::renderGpuSection()
             ImGui::TableNextColumn();
             const ImVec4 gpuMemColor = theme.scheme().gpuMemory;
             const std::string memStr = std::format("{} / {} ({:.1f}%%)",
-                                                   UI::Format::formatBytes(snap.memoryUsedBytes),
-                                                   UI::Format::formatBytes(snap.memoryTotalBytes),
+                                                   UI::Format::formatBytes(static_cast<double>(snap.memoryUsedBytes)),
+                                                   UI::Format::formatBytes(static_cast<double>(snap.memoryTotalBytes)),
                                                    smoothed.memoryPercent);
             ImGui::TextColored(gpuMemColor, "%s", memStr.c_str());
